Remove old filters.py and move new filters.py to pyslyde/util
